+++ db/migrations/sqlite3/0001_census3.sql
@@ -1,11 +1,11 @@
 -- +goose Up
 CREATE TABLE Strategies (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     predicate TEXT NOT NULL
 );
 
 CREATE TABLE TokenTypes (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     type_name TEXT NOT NULL UNIQUE
 );
 
@@ -13,17 +13,17 @@ CREATE TABLE Tokens (
     id BLOB PRIMARY KEY NOT NULL,
     name TEXT,
     symbol TEXT,
-    decimals INTEGER,
+    decimals BIGINT,
     total_supply BLOB,
-    creation_block INTEGER NOT NULL,
-    type_id INTEGER NOT NULL,
+    creation_block BIGINT NOT NULL,
+    type_id BIGINT NOT NULL,
     FOREIGN KEY (type_id) REFERENCES TokenTypes(id) ON DELETE CASCADE
 );
 CREATE INDEX idx_tokens_type_id ON Tokens(type_id);
 
 CREATE TABLE Censuses (
-    id INTEGER PRIMARY KEY,
-    strategy_id INTEGER NOT NULL,
+    id BIGINT PRIMARY KEY,
+    strategy_id BIGINT NOT NULL,
     merkle_root BLOB NOT NULL UNIQUE,
     uri TEXT UNIQUE,
     FOREIGN KEY (strategy_id) REFERENCES Strategies(id) ON DELETE CASCADE
@@ -31,7 +31,7 @@ CREATE TABLE Censuses (
 CREATE INDEX idx_censuses_strategy_id ON Censuses(strategy_id);
 
 CREATE TABLE Blocks (
-    id INTEGER PRIMARY KEY NOT NULL,
+    id BIGINT PRIMARY KEY NOT NULL,
     timestamp TEXT NOT NULL UNIQUE,
     root_hash BLOB NOT NULL UNIQUE
 );
@@ -44,7 +44,7 @@ CREATE TABLE TokenHolders (
     token_id BLOB NOT NULL,
     holder_id BLOB NOT NULL,
     balance BLOB NOT NULL,
-    block_id INTEGER NOT NULL,
+    block_id BIGINT NOT NULL,
     PRIMARY KEY (token_id, holder_id, block_id),
     FOREIGN KEY (token_id) REFERENCES Tokens(id) ON DELETE CASCADE,
     FOREIGN KEY (holder_id) REFERENCES Holders(id) ON DELETE CASCADE,
@@ -55,7 +55,7 @@ CREATE INDEX idx_tokenholders_holder_id ON TokenHolders(holder_id);
 CREATE INDEX idx_tokenholders_block_id ON TokenHolders(block_id);
 
 CREATE TABLE StrategyTokens (
-    strategy_id INTEGER NOT NULL,
+    strategy_id BIGINT NOT NULL,
     token_id BLOB NOT NULL,
     min_balance BLOB NOT NULL,
     method_hash BLOB NOT NULL,
@@ -67,8 +67,8 @@ CREATE INDEX idx_strategytokens_strategy_id ON StrategyTokens(strategy_id);
 CREATE INDEX idx_strategytokens_token_id ON StrategyTokens(token_id);
 
 CREATE TABLE CensusBlocks (
-    census_id INTEGER NOT NULL,
-    block_id INTEGER NOT NULL,
+    census_id BIGINT NOT NULL,
+    block_id BIGINT NOT NULL,
     PRIMARY KEY (census_id, block_id),
     FOREIGN KEY (census_id) REFERENCES Censuses(id) ON DELETE CASCADE,
     FOREIGN KEY (block_id) REFERENCES Blocks(id) ON DELETE CASCADE
